move `GuessBaseImageIndex` to reduce dependencies
