+++ pkg/fanal/artifact/image/image.go
@@ -17,7 +17,7 @@ import (
 	"github.com/aquasecurity/trivy/pkg/fanal/analyzer"
 	"github.com/aquasecurity/trivy/pkg/fanal/artifact"
 	"github.com/aquasecurity/trivy/pkg/fanal/handler"
-	"github.com/aquasecurity/trivy/pkg/fanal/image"
+	imageutils "github.com/aquasecurity/trivy/pkg/fanal/image/utils"
 	"github.com/aquasecurity/trivy/pkg/fanal/types"
 	"github.com/aquasecurity/trivy/pkg/fanal/walker"
 	"github.com/aquasecurity/trivy/pkg/log"
@@ -391,7 +391,7 @@ func (a Artifact) guessBaseLayers(diffIDs []string, configFile *v1.ConfigFile) [
 		return nil
 	}
 
-	baseImageIndex := image.GuessBaseImageIndex(configFile.History)
+	baseImageIndex := imageutils.GuessBaseImageIndex(configFile.History)
 
 	// Diff IDs don't include empty layers, so the index is different from histories
 	var diffIDIndex int
